pyproject.toml
[project]
name = "snekquiz"
version = "0.1.1"
description = "A minimalist MCQ quiz web application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "aiosqlite>=0.21.0",
    "ldap3>=2.9",
]

[project.scripts]
snekquiz = "snekquiz:main"

[build-system]
requires = ["uv_build>=0.9.25,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # ruff-specific rules
    "S",      # flake8-bandit (security)
    "A",      # flake8-builtins
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T20",    # flake8-print
    "PT",     # flake8-pytest-style
    "LOG",    # flake8-logging
    "ASYNC",  # flake8-async
]
ignore = [
    "S104",   # binding to all interfaces
    "S105",   # hardcoded password in config defaults
]

[tool.ruff.lint.isort]
known-first-party = ["snekquiz"]


[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "prek>=0.3.3",
    "ruff>=0.15.2",
    "ty>=0.0.18",
]
